+++ example/google_example.dart
@@ -13,7 +13,7 @@ void main() async {
 
     //Pick an English Voice
     final voiceGoogle = voicesResponseGoogle.voices
-        .where((element) => element.locale.startsWith("en-"))
+        .where((element) => element.locale.code.startsWith("en-"))
         .toList(growable: false)
         .first;