+++ example/universal_example.dart
@@ -18,7 +18,7 @@ void main() async {
 
     //Pick an English Voice
     final voiceUniversal = voicesResponseUniversal.voices
-        .where((element) => element.locale.startsWith("en-"))
+        .where((element) => element.locale.code.startsWith("en-"))
         .toList(growable: false)
         .first;
 
@@ -53,7 +53,7 @@ void main() async {
 
     //Pick an English Voice
     final voiceUniversal2 = voicesResponseUniversal2.voices
-        .where((element) => element.locale.startsWith("en-"))
+        .where((element) => element.locale.code.startsWith("en-"))
         .toList(growable: false)
         .first;