+++ lib/src/google/voices/voice_model.dart
@@ -4,7 +4,7 @@ import 'package:cloud_text_to_speech/src/common/utils/helpers.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:locale_names/locale_names.dart';
 
-import '../../universal/voices/voice_model.dart';
+import '../../../cloud_text_to_speech.dart';
 
 part 'voice_model.g.dart';
 
@@ -19,14 +19,7 @@ class VoiceGoogle extends VoiceUniversal {
   @JsonKey(name: "ssmlGender", fromJson: _toGender)
   String gender;
   @JsonKey(name: "languageCodes", fromJson: _toLocale, includeToJson: false)
-  String locale;
-  @JsonKey(name: "languageCodes", fromJson: _toLocaleName, includeToJson: false)
-  String localeName;
-  @JsonKey(
-      name: "languageCodes",
-      fromJson: _toNativeLocaleName,
-      includeToJson: false)
-  String nativeLocaleName;
+  VoiceLocale locale;
   @JsonKey(
       name: "naturalSampleRateHertz",
       fromJson: _toSampleRateHertz,
@@ -42,8 +35,6 @@ class VoiceGoogle extends VoiceUniversal {
     required this.nativeName,
     required this.gender,
     required this.locale,
-    required this.localeName,
-    required this.nativeLocaleName,
     required this.sampleRateHertz,
   }) : super(
             code: code,
@@ -52,58 +43,42 @@ class VoiceGoogle extends VoiceUniversal {
             nativeName: nativeName,
             gender: gender,
             locale: locale,
-            localeName: locale,
-            nativeLocaleName: nativeLocaleName,
             sampleRateHertz: sampleRateHertz);
 
   factory VoiceGoogle.fromJson(Map<String, dynamic> json) =>
       _$VoiceGoogleFromJson(json);
+
+  static String _toVoiceType(String name) {
+    List<String> nameSegments = name.split('-');
+    return nameSegments[2];
+  }
 
   static String _toGender(String ssmlGender) {
     final lowercase = ssmlGender.toLowerCase();
     return lowercase[0].toUpperCase() + lowercase.substring(1);
   }
 
-  static String _toLocale(List<dynamic> languageCodes) {
+  static VoiceLocale _toLocale(List<dynamic> languageCodes) {
     if (languageCodes.isNotEmpty) {
-      return languageCodes[0] as String;
+      List<String> localeSegments = (languageCodes[0] as String).split('-');
+
+      Locale localeObj = Helpers.segmentsToLocale(localeSegments);
+
+      return VoiceLocale(
+          code: localeSegments.join('-'),
+          name: Helpers.formatLanguageCountry(localeObj.defaultDisplayLanguage,
+              localeObj.defaultDisplayCountry),
+          nativeName: Helpers.formatLanguageCountry(
+              localeObj.nativeDisplayLanguage, localeObj.nativeDisplayCountry),
+          languageCode: localeObj.languageCode,
+          languageName: localeObj.defaultDisplayLanguage,
+          nativeLanguageName: localeObj.nativeDisplayLanguage,
+          countryCode: localeObj.countryCode,
+          countryName: localeObj.defaultDisplayCountry,
+          nativeCountryName: localeObj.nativeDisplayCountry,
+          scriptCode: localeObj.scriptCode);
     }
-    return '';
-  }
-
-  static String _toLocaleName(List<dynamic> languageCodes) {
-    if (languageCodes.isNotEmpty) {
-      List<String> languageCountrySegments =
-          (languageCodes[0] as String).split('-');
-
-      Locale localeObj = Locale.fromSubtags(
-          languageCode: languageCountrySegments[0],
-          countryCode: languageCountrySegments[1]);
-
-      return Helpers.formatLanguageCountry(
-          localeObj.defaultDisplayLanguage, localeObj.defaultDisplayCountry);
-    }
-    return '';
-  }
-
-  static String _toNativeLocaleName(List<dynamic> languageCodes) {
-    if (languageCodes.isNotEmpty) {
-      List<String> languageCountrySegments =
-          (languageCodes[0] as String).split('-');
-
-      Locale localeObj = Locale.fromSubtags(
-          languageCode: languageCountrySegments[0],
-          countryCode: languageCountrySegments[1]);
-
-      return Helpers.formatLanguageCountry(
-          localeObj.nativeDisplayLanguage, localeObj.nativeDisplayCountry);
-    }
-    return '';
-  }
-
-  static String _toVoiceType(String name) {
-    List<String> nameSegments = name.split('-');
-    return nameSegments[2];
+    return VoiceLocale.code(code: '');
   }
 
   static String _toSampleRateHertz(int naturalSampleRateHertz) {