+++ lib/src/google/voices/voice_model.g.dart
@@ -13,9 +13,6 @@ VoiceGoogle _$VoiceGoogleFromJson(Map<String, dynamic> json) => VoiceGoogle(
       nativeName: json['name'] as String,
       gender: VoiceGoogle._toGender(json['ssmlGender'] as String),
       locale: VoiceGoogle._toLocale(json['languageCodes'] as List),
-      localeName: VoiceGoogle._toLocaleName(json['languageCodes'] as List),
-      nativeLocaleName:
-          VoiceGoogle._toNativeLocaleName(json['languageCodes'] as List),
       sampleRateHertz:
           VoiceGoogle._toSampleRateHertz(json['naturalSampleRateHertz'] as int),
     );