+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 * Add Google and Microsoft TTS
+
+## 1.1.0
+
+* Separate locale from voice
+* Sort by locale and gender